Add flag that indicates if scalars are quoted
This is useful for preserving types: YAML allows for strings
to be unquoted. Thus an unquoted string "false" can not be
distinguished from the bool value false.

The usual way to resolve this is to explicitly quote
ambiguative strings. Thus a string "2.0" would be quoted,
while the number would not be quoted.

Multi line strings are also considered to be quoted.